+++ demo.py
@@ -1,0 +1,6 @@
+from lihil import Lihil
+
+lhl = Lihil()
+
+
+lhl.run(__file__)