+++ lihil/config/parser.py
@@ -5,15 +5,8 @@ from msgspec.structs import FieldInfo, fields
 from typing_extensions import Doc
 
 from lihil.config.app_config import ConfigBase
-from lihil.interface import MISSING, P, Record, StrDict, UnsetType, _Missed, is_provided
+from lihil.interface import MISSING, Record, StrDict, UnsetType, _Missed, is_provided
 from lihil.utils.typing import get_origin_pro, is_union_type, lexient_issubclass
-
-
-def get_thread_cnt() -> int:
-    import os
-
-    default_max = os.cpu_count() or 1
-    return default_max
 
 
 def format_nested_dict(flat_dict: StrDict) -> StrDict:
@@ -89,9 +82,7 @@ def generate_parser_actions(
 
     for field_info in cls_fields:
         field_name = field_info.encode_name
-        field_default: Any = (
-            MISSING  # if field_type is not bool else field_info.default
-        )
+        field_default: Any = MISSING
 
         full_field_name = f"{prefix}.{field_name}" if prefix else field_name
         arg_name = f"--{full_field_name}"