+++ CHANGELOG.md
@@ -1078,3 +1078,37 @@ async def test_throttling():
 ### Refactor
 
 - [x] remove `registry` and `listeners` from `Route`
+
+- [x] merge props from `Route` and `Endpoint` instead of update
+
+- [x] deduplicate plugins
+
+```python
+async def test_route_merge_endpoint_plugin():
+    called: list[str] = []
+
+    async def dummy_plugin(*args):
+        called.append("plugin called")
+
+    route = Route(props=EndpointProps(plugins=[dummy_plugin]))
+
+    async def dummy_handler(): ...
+
+    route.get(dummy_handler, plugins=[dummy_plugin])
+
+    await route.setup()
+
+    ep = route.get_endpoint(dummy_handler)
+
+    # merged plugin from route and endpoint so we have two plugins
+    assert ep.props.plugins == [dummy_plugin, dummy_plugin]
+
+    # but only one plugin called since we deduplicate using id(plugin)
+    assert called == ["plugin called"]
+```
+
+-[x] `Lihil(max_thread_workers=x)` would set max thread workers to x, this would be used in `Graph` and `Endpoint` for executing sync function in thread workers.
+
+### Fixes
+
+- [x] fix a bug where if lihil will not read config from cli upon initialization.